getHours()

+++ src/main/java/com/mtv/erp/model/Hours.java
@@ -7,10 +7,10 @@ public class Hours {
     private int id;
     private User user;
     private LocalDate date;
-    private int hours;
+    private float hours;
     private boolean saved;
 
-    public Hours(int id, User user, LocalDate date, int hours, boolean saved) {
+    public Hours(int id, User user, LocalDate date, float hours, boolean saved) {
         this.id = id;
         this.user = user;
         this.date = date;
@@ -18,20 +18,20 @@ public class Hours {
         this.saved = saved;
     }
 
-    public Hours(int id, User user, LocalDate date, int hours) {
+    public Hours(int id, User user, LocalDate date, float hours) {
         this(id, user, date, hours, false);
     }
 
     public Hours(){
     }
 
-    public Hours(int hours, LocalDate date, User user){
+    public Hours(float hours, LocalDate date, User user){
         setHours(hours);
         setDate(date);
         setUser(user);
     }
 
-    public Hours(int id, int hours){
+    public Hours(int id, float hours){
         this(id, null, null, hours);
     }
 
@@ -59,11 +59,11 @@ public class Hours {
         this.date = date;
     }
 
-    public int getHours() {
+    public float getHours() {
         return hours;
     }
 
-    public void setHours(int hours) {
+    public void setHours(float hours) {
         this.hours = hours;
     }
